libtock: app_state: remove warning on rv64
The app state won't work on RISCV-64

+++ libtock/storage/app_state.c
@@ -58,6 +58,6 @@ returncode_t libtock_app_state_save(libtock_app_state_callback cb) {
   err = libtock_app_state_set_upcall(app_state_upcall, (void*) cb);
   if (err != RETURNCODE_SUCCESS) return err;
 
-  err = libtock_app_state_command_save((uint32_t) _app_state_flash_pointer);
+  err = libtock_app_state_command_save((uintptr_t) _app_state_flash_pointer);
   return err;
 }
